requirements.txt
openai-agents # or `uv add openai-agents`, etc
python-dotenv